+++ classes/player/config.lua
@@ -9,6 +9,7 @@ local playerConfig = {
 	crosshair_max_distance = 60,
 	crosshair_radius = 6,
 	crosshair_line = 10,
+	hand_distance = 10
 }
 
 return playerConfig